Big Update
Added Insurance selection panel if dealer showing ace. Update chip
counts working for tester. Implemented Insurance action. Made
UserController static with access to chipUpdateCount method. And more.

+++ Code/CMSC495-BlackJack/src/com/blackjack/client/service/UserService.java
@@ -19,7 +19,7 @@ public interface UserService extends RemoteService {
 	
 	public CreateAccountEvent createAccount(String username, String password, String email);
 
-	public UpdateChipEvent updateChipCount(String userID, int amount);
+	public UpdateChipEvent updateChipCount(String userID, float amount);
 
 	public ResetPasswordEvent resetPassword(String emailAddress);
